chore: refactor logic in hook
