+++ sample-apps/react-native/ExpoTikTokApp/components/common/annotated-text/MentionPreview.tsx
@@ -1,4 +1,4 @@
-import { StyleSheet, Text, TextInputProps } from 'react-native';
+import { StyleSheet, Text } from 'react-native';
 import React from 'react';
 import { useStableCallback } from '@/hooks/useStableCallback';
